+++ src/assets/index.scss
@@ -1,7 +1,11 @@
 @import '~bootstrap/scss/bootstrap.scss';
-@import './App.scss';
+@import './home.scss';
 
 body {
+  box-sizing: border-box;
+  display: flex;
+  justify-content: center;
+  align-items: center;
   margin: 0;
   font-family:
     -apple-system,
@@ -17,14 +21,4 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family:
-    source-code-pro,
-    Menlo,
-    Monaco,
-    Consolas,
-    'Courier New',
-    monospace;
 }
